Add DeleteEntry query
