+++ schema/schema.graphql
@@ -133,6 +133,7 @@ type Mutation {
   completeUserSignup(input: SignedUpUser!): User!
   createEntry(input: NewEntry!): Entry!
   updateEntry(id: ID!, input: ExistingEntry!, date: String!): Entry!
+  deleteEntry(id: ID!): Entry!
   createEditor(input: NewEditor!): Editor!
   createSubscription(input: NewSubscription!): StripeSubscription!
   cancelSubscription(id: ID!): String!
